Created new shell to prevent breaking other tests
